refactor: move interfaces to types folder and update imports
